Compact Cie and Fde (64 -> 48, 88 -> 56 bytes)
Both DWARF unwind records copied their bytes into an owned Vec<u8> (24
bytes of header plus the copy) and indexed their object, CIE and
subsection with usize; the FDE's LSDA was an Option<(usize, u32)>.
Borrow the record bytes from the object's __eh_frame instead - the
parser's processed copy of the section (subtraction pairs pre-applied)
is leaked once per object, like its section headers, and the records
slice it; mold-rust's CieRecord borrows its contents
the same way - and store every index as u32 (the LSDA as an
Option<(u32, u32)>). Cie is 48 bytes, Fde 56, asserted at compile
time; mold-rust's FdeRecord is 16 because it derives nearly everything
from the CIE and the section, which is a larger change than the
records' current readers warrant.

A debug clang link on arm64 has no __eh_frame at all (compact unwind
covers every function), so this is for x86-64 and DWARF-unwind-heavy
inputs. 322 integration tests pass; the debug clang, nu and mold links
run.

+++ src/dead_strip.rs
@@ -135,9 +135,9 @@ pub fn dead_strip<E: Arch>(ctx: &mut Context<E>) {
             let mut personality = rec.personality();
             if let Some(fde) = rec.fde() {
                 if let Some((lsda, _)) = ctx.fdes[fde].lsda {
-                    out.push(lsda);
+                    out.push(lsda as usize);
                 }
-                personality = personality.or(ctx.cies[ctx.fdes[fde].cie].personality);
+                personality = personality.or(ctx.cies[ctx.fdes[fde].cie as usize].personality);
             }
             if let Some(p) = personality {
                 if let Some(isec) = ctx.symtab[p].isec().map(|i| i as usize) {
@@ -196,10 +196,10 @@ pub fn dead_strip<E: Arch>(ctx: &mut Context<E>) {
                 let mut personality = rec.personality();
                 if let Some(fde) = rec.fde() {
                     if let Some((lsda, _)) = gc.ctx.fdes[fde].lsda {
-                        targets.push(lsda);
+                        targets.push(lsda as usize);
                     }
                     personality =
-                        personality.or(gc.ctx.cies[gc.ctx.fdes[fde].cie].personality);
+                        personality.or(gc.ctx.cies[gc.ctx.fdes[fde].cie as usize].personality);
                 }
                 if let Some(p) = personality {
                     if let Some(isec) = gc.ctx.symtab[p].isec().map(|i| i as usize) {
@@ -273,7 +273,7 @@ pub fn dead_strip<E: Arch>(ctx: &mut Context<E>) {
     let mut kept_fdes = Vec::new();
     let fdes = std::mem::take(&mut ctx.fdes);
     for (i, fde) in fdes.into_iter().enumerate() {
-        if live[fde.isec] {
+        if live[fde.isec as usize] {
             fde_map[i] = kept_fdes.len();
             kept_fdes.push(fde);
         }
